Refactoring file names
