+++ app_guizero/app_logs.py
@@ -22,8 +22,8 @@ from tkinter import filedialog
 from guizero import Window, Text, TextBox, ButtonGroup, PushButton, MenuBar, info, warn
 
 import app_modules.app_logger as app_logger
-import app_modules.app_sensor_commands as app_sensor_commands
-from app_modules.app_useful import no_ip_selected_message
+import app_modules.sensor_commands as app_sensor_commands
+from app_modules.app_variables import no_ip_selected_message
 
 
 class CreateSensorLogsWindow:
