+++ app_modules/graphing_live.py
@@ -21,9 +21,9 @@ from datetime import datetime
 from matplotlib import pyplot, animation, style
 
 import app_modules.app_logger as app_logger
-import app_modules.app_sensor_commands as app_sensor_commands
-from app_modules.app_graph import CreateMeasurementsTypes, CreateSQLColumnsReadable, CreateSQLColumnNames
-from app_modules.app_useful import convert_minutes_string
+import app_modules.sensor_commands as app_sensor_commands
+from app_modules.graphing import CreateMeasurementsTypes, CreateSQLColumnsReadable, CreateSQLColumnNames
+from app_modules.app_variables import convert_minutes_string
 
 style.use("dark_background")
