+++ app_modules/graphing_offline_extras.py
