+++ app_modules/reports.py
@@ -21,8 +21,8 @@ from threading import Thread
 from time import strftime
 
 import app_modules.app_logger as app_logger
-import app_modules.app_sensor_commands as app_sensor_commands
-import app_modules.app_useful as useful
+import app_modules.sensor_commands as app_sensor_commands
+import app_modules.app_variables as useful
 
 network_get_commands = app_sensor_commands.CreateNetworkGetCommands()
